requirements.txt
Django==2.2.12
djangorestframework==3.11.0
psycopg2==2.8.5
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
django-widget-tweaks == 1.4.8
docx2pdf==0.1.7
django-ckeditor==5.9.0
WeasyPrint==51
python-docx==0.8.10
mammoth==1.4.11
gunicorn==20.0.4
